Progress on CSS by doing CSS Flag of Laos
NOTE: I really had a little bit of hard time doing the project as I need to analyze how the CSS selector works with the Parent - Descendant relation works. 
I was able to grasp how the position: absolute and relative work, though the solution and my work has a different code on them, the outcome was the same. I still need to analyze how to improve my code vs the solution they provided.